+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- feat: Add helpful aggregate functions to the Query Builder.
+
 ## [3.0.2] - 2021-02-04
 
 ### Changed